+++ app/client/src/components/BottomBar/components.ts
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+import { Layers } from "constants/Layers";
+
+export const Container = styled.div`
+  width: 100%;
+  height: ${(props) => props.theme.bottomBarHeight};
+  display: flex;
+  position: fixed;
+  justify-content: space-between;
+  background-color: ${(props) => props.theme.colors.editorBottomBar.background};
+  z-index: ${Layers.bottomBar};
+  border-top: solid 1px var(--ads-v2-color-border);
+`;
+
+export const Wrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
