+++ app/client/src/pages/Editor/EditorHeader.tsx
@@ -1,18 +1,7 @@
-import React, {
-  useCallback,
-  useEffect,
-  useState,
-  lazy,
-  Suspense,
-  useContext,
-} from "react";
-import styled, { ThemeProvider } from "styled-components";
-import classNames from "classnames";
-import { APPLICATIONS_URL } from "constants/routes";
+import React, { useCallback, useEffect, useState, useContext } from "react";
+import { ThemeProvider } from "styled-components";
 import AppInviteUsersForm from "pages/workspace/AppInviteUsersForm";
 import AnalyticsUtil from "utils/AnalyticsUtil";
-import AppsmithLogo from "assets/images/appsmith_logo_square.png";
-import { Link } from "react-router-dom";
 import {
   getApplicationLastDeployedAt,
   getCurrentApplicationId,
@@ -21,7 +10,6 @@ import {
   previewModeSelector,
 } from "selectors/editorSelectors";
 import {
-  getAllUsers,
   getCurrentAppWorkspace,
   getCurrentWorkspaceId,
 } from "@appsmith/selectors/workspaceSelectors";
@@ -38,15 +26,9 @@ import {
   getCurrentApplication,
 } from "@appsmith/selectors/applicationSelectors";
 import EditorAppName from "./EditorAppName";
-import { getCurrentUser } from "selectors/usersSelectors";
-import {
-  EditInteractionKind,
-  SavingState,
-  getTypographyByKey,
-} from "design-system-old";
+import { EditInteractionKind, SavingState } from "design-system-old";
 import {
   Button,
-  Icon,
   Tooltip,
   Modal,
   ModalHeader,
@@ -57,11 +39,8 @@ import {
   Tab,
   TabPanel,
 } from "design-system";
-import { Profile } from "pages/common/ProfileImage";
-import HelpBar from "components/editorComponents/GlobalSearch/HelpBar";
 import { getTheme, ThemeMode } from "selectors/themeSelectors";
 import ToggleModeButton from "pages/Editor/ToggleModeButton";
-import { snipingModeSelector } from "selectors/editorSelectors";
 import { showConnectGitModal } from "actions/gitSyncActions";
 import RealtimeAppEditors from "./RealtimeAppEditors";
 import { EditorSaveIndicator } from "./EditorSaveIndicator";
@@ -69,32 +48,19 @@ import {
   adaptiveSignpostingEnabled,
   selectFeatureFlags,
 } from "@appsmith/selectors/featureFlagsSelectors";
-import { retryPromise } from "utils/AppsmithUtils";
 import { fetchUsersForWorkspace } from "@appsmith/actions/workspaceActions";
 
 import { getIsGitConnected } from "selectors/gitSyncSelectors";
 import {
-  CLOSE_ENTITY_EXPLORER_MESSAGE,
   createMessage,
   DEPLOY_BUTTON_TOOLTIP,
   DEPLOY_MENU_OPTION,
-  EDITOR_HEADER,
   INVITE_TAB,
   IN_APP_EMBED_SETTING,
-  LOCK_ENTITY_EXPLORER_MESSAGE,
-  LOGO_TOOLTIP,
   RENAME_APPLICATION_TOOLTIP,
-  SHARE_BUTTON_TOOLTIP,
-  SHARE_BUTTON_TOOLTIP_WITH_USER,
   COMMUNITY_TEMPLATES,
   APPLICATION_INVITE,
 } from "@appsmith/constants/messages";
-import { getExplorerPinned } from "selectors/explorerSelector";
-import {
-  setExplorerActiveAction,
-  setExplorerPinnedAction,
-} from "actions/explorerActions";
-import { modText } from "utils/helpers";
 import Boxed from "./GuidedTour/Boxed";
 import EndTour from "./GuidedTour/EndTour";
 import { GUIDED_TOUR_STEPS } from "./GuidedTour/constants";
@@ -118,126 +84,21 @@ import PublishCommunityTemplateModal from "./CommunityTemplates/Modals/PublishCo
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
 import { getEmbedSnippetForm } from "@appsmith/utils/BusinessFeatures/privateEmbedHelpers";
+import { HeaderSection, HeaderWrapper } from "./commons/EditorHeaderComponents";
+import { LockEntityExplorer } from "./commons/LockEntityExplorer";
+import { Omnibar } from "./commons/Omnibar";
+import { EditorShareButton } from "./EditorShareButton";
+import { HelperBarInHeader } from "./HelpBarInHeader";
+import { AppsmithLink } from "./AppsmithLink";
 
 const { cloudHosting } = getAppsmithConfigs();
-
-const HeaderWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  background-color: var(--ads-v2-color-bg);
-  flex-direction: row;
-  box-shadow: none;
-  border-bottom: 1px solid var(--ads-v2-color-border);
-  height: ${(props) => props.theme.smallHeaderHeight};
-  & .editable-application-name {
-    ${getTypographyByKey("h4")}
-    color: ${(props) => props.theme.colors.header.appName};
-  }
-  & ${Profile} {
-    width: 24px;
-    height: 24px;
-  }
-
-  @media only screen and (max-width: 900px) {
-    & .help-bar {
-      display: none;
-    }
-  }
-
-  @media only screen and (max-width: 700px) {
-    & .app-realtime-editors {
-      display: none;
-    }
-  }
-`;
-
-const HeaderSection = styled.div`
-  position: relative;
-  display: flex;
-  flex: 1;
-  overflow: visible;
-  align-items: center;
-  :nth-child(1) {
-    justify-content: flex-start;
-    max-width: 30%;
-  }
-  :nth-child(2) {
-    justify-content: center;
-  }
-  :nth-child(3) {
-    justify-content: flex-end;
-  }
-`;
-
-const AppsmithLink = styled((props) => {
-  // we are removing non input related props before passing them in the components
-  // eslint-disable @typescript-eslint/no-unused-vars
-  return <Link {...props} />;
-})`
-  height: 24px;
-  min-width: 24px;
-  width: 24px;
-  display: inline-block;
-  img {
-    min-width: 24px;
-    width: 24px;
-    height: 24px;
-  }
-`;
-
-const BindingBanner = styled.div`
-  position: fixed;
-  width: 199px;
-  height: 36px;
-  left: 50%;
-  top: ${(props) => props.theme.smallHeaderHeight};
-  transform: translate(-50%, 0);
-  text-align: center;
-  background: var(--ads-v2-color-fg-information);
-  color: var(--ads-v2-color-white);
-  border-radius: var(--ads-v2-border-radius);
-  font-weight: 500;
-  font-size: 15px;
-  line-height: 20px;
-  /* Depth: 01 */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-shadow: var(--ads-v2-shadow-popovers);
-  z-index: 9999;
-`;
-
-const SidebarNavButton = styled(Button)`
-  .ads-v2-button__content {
-    padding: 0;
-  }
-  .group {
-    height: 36px;
-    width: 36px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-`;
-
-const GlobalSearch = lazy(async () => {
-  return retryPromise(
-    async () =>
-      import(
-        /* webpackChunkName: "global-search" */ "components/editorComponents/GlobalSearch"
-      ),
-  );
-});
 
 const theme = getTheme(ThemeMode.LIGHT);
 
 export function EditorHeader() {
   const [activeTab, setActiveTab] = useState("invite");
   const dispatch = useDispatch();
-  const isSnipingMode = useSelector(snipingModeSelector);
   const isSavingName = useSelector(getIsSavingAppName);
-  const pinned = useSelector(getExplorerPinned);
   const isGitConnected = useSelector(getIsGitConnected);
   const isErroredSavingName = useSelector(getIsErroredSavingAppName);
   const applicationList = useSelector(getApplicationList);
@@ -249,8 +110,6 @@ export function EditorHeader() {
   const currentApplication = useSelector(getCurrentApplication);
   const isPublishing = useSelector(getIsPublishingApplication);
   const pageId = useSelector(getCurrentPageId) as string;
-  const sharedUserList = useSelector(getAllUsers);
-  const currentUser = useSelector(getCurrentUser);
   const featureFlags = useSelector(selectFeatureFlags);
 
   const deployLink = useHref(viewerURL, { pageId });
@@ -338,29 +197,12 @@ export function EditorHeader() {
     [dispatch, handlePublish],
   );
 
-  /**
-   * on hovering the menu, make the explorer active
-   */
-  const onMenuHover = useCallback(() => {
-    dispatch(setExplorerActiveAction(true));
-  }, [setExplorerActiveAction]);
-
-  /**
-   * toggles the pinned state of sidebar
-   */
-  const onPin = useCallback(() => {
-    dispatch(setExplorerPinnedAction(!pinned));
-  }, [pinned, dispatch, setExplorerPinnedAction]);
-
   //Fetch all users for the application to show the share button tooltip
   useEffect(() => {
     if (workspaceId) {
       dispatch(fetchUsersForWorkspace(workspaceId));
     }
   }, [workspaceId]);
-  const filteredSharedUserList = sharedUserList.filter(
-    (user) => user.username !== currentUser?.username,
-  );
 
   const {
     isOpened: isWalkthroughOpened,
@@ -412,74 +254,10 @@ export function EditorHeader() {
       >
         <HeaderSection className="space-x-2">
           {!signpostingEnabled && (
-            <Tooltip
-              content={
-                <div className="flex items-center justify-between">
-                  <span>
-                    {!pinned
-                      ? createMessage(LOCK_ENTITY_EXPLORER_MESSAGE)
-                      : createMessage(CLOSE_ENTITY_EXPLORER_MESSAGE)}
-                  </span>
-                  <span className="ml-4">{modText()} /</span>
-                </div>
-              }
-              placement="bottomLeft"
-            >
-              <SidebarNavButton
-                className={classNames({
-                  "transition-all transform duration-400": true,
-                  "-translate-x-full opacity-0": isPreviewingApp,
-                  "translate-x-0 opacity-100": !isPreviewingApp,
-                })}
-                data-testid="sidebar-nav-button"
-                kind="tertiary"
-                onClick={onPin}
-                size="md"
-              >
-                <div
-                  className="relative t--pin-entity-explorer group"
-                  onMouseEnter={onMenuHover}
-                >
-                  <Icon
-                    className="absolute transition-opacity group-hover:opacity-0"
-                    name="hamburger"
-                    size="md"
-                  />
-                  {pinned && (
-                    <Icon
-                      className="absolute transition-opacity opacity-0 group-hover:opacity-100"
-                      name="menu-fold"
-                      onClick={onPin}
-                      size="md"
-                    />
-                  )}
-                  {!pinned && (
-                    <Icon
-                      className="absolute transition-opacity opacity-0 group-hover:opacity-100"
-                      name="menu-unfold"
-                      onClick={onPin}
-                      size="md"
-                    />
-                  )}
-                </div>
-              </SidebarNavButton>
-            </Tooltip>
+            <LockEntityExplorer isPreviewingApp={isPreviewingApp} />
           )}
 
-          <Tooltip content={createMessage(LOGO_TOOLTIP)} placement="bottomLeft">
-            <AppsmithLink
-              className={classNames({
-                "ml-2": signpostingEnabled,
-              })}
-              to={APPLICATIONS_URL}
-            >
-              <img
-                alt="Appsmith logo"
-                className="t--appsmith-logo"
-                src={AppsmithLogo}
-              />
-            </AppsmithLink>
-          </Tooltip>
+          <AppsmithLink />
 
           <Tooltip
             content={createMessage(RENAME_APPLICATION_TOOLTIP)}
@@ -514,16 +292,9 @@ export function EditorHeader() {
           </Tooltip>
           <EditorSaveIndicator />
         </HeaderSection>
-        <HeaderSection
-          className={classNames({
-            "-translate-y-full opacity-0": isPreviewMode,
-            "translate-y-0 opacity-100": !isPreviewMode,
-            "transition-all transform duration-400": true,
-            "help-bar": "true",
-          })}
-        >
-          <HelpBar />
-        </HeaderSection>
+
+        <HelperBarInHeader />
+
         <HeaderSection className="gap-x-1">
           <Boxed
             alternative={<EndTour />}
@@ -531,30 +302,7 @@ export function EditorHeader() {
           >
             <RealtimeAppEditors applicationId={applicationId} />
             <ToggleModeButton />
-            {applicationId && (
-              <Tooltip
-                content={
-                  filteredSharedUserList.length
-                    ? createMessage(
-                        SHARE_BUTTON_TOOLTIP_WITH_USER(
-                          filteredSharedUserList.length,
-                        ),
-                      )
-                    : createMessage(SHARE_BUTTON_TOOLTIP)
-                }
-                placement="bottom"
-              >
-                <Button
-                  className="t--application-share-btn"
-                  kind="tertiary"
-                  onClick={() => setShowModal(true)}
-                  size="md"
-                  startIcon="share-line"
-                >
-                  {createMessage(EDITOR_HEADER.share)}
-                </Button>
-              </Tooltip>
-            )}
+            {applicationId && <EditorShareButton setShowModal={setShowModal} />}
             <Modal
               onOpenChange={(isOpen) => setShowModal(isOpen)}
               open={showModal}
@@ -640,14 +388,7 @@ export function EditorHeader() {
             </div>
           </Boxed>
         </HeaderSection>
-        <Suspense fallback={<span />}>
-          <GlobalSearch />
-        </Suspense>
-        {isSnipingMode && (
-          <BindingBanner className="t--sniping-mode-banner">
-            Select a widget to bind
-          </BindingBanner>
-        )}
+        <Omnibar />
       </HeaderWrapper>
     </ThemeProvider>
   );
