+++ app/client/src/pages/Editor/Explorer/Datasources.tsx
@@ -1,20 +1,6 @@
-import React, { useCallback, useContext } from "react";
-import {
-  useAppWideAndOtherDatasource,
-  useDatasourceSuggestions,
-} from "./hooks";
-import type { Datasource } from "entities/Datasource";
-import ExplorerDatasourceEntity from "./Datasources/DatasourceEntity";
+import React from "react";
 import { useSelector } from "react-redux";
-import {
-  getCurrentApplicationId,
-  getCurrentPageId,
-} from "selectors/editorSelectors";
-import { getPlugins } from "@appsmith/selectors/entitiesSelector";
-import { keyBy } from "lodash";
 import Entity from "./Entity";
-import history from "utils/history";
-import { INTEGRATION_TABS } from "constants/routes";
 import {
   ADD_DATASOURCE_BUTTON,
   createMessage,
@@ -23,45 +9,53 @@ import {
   EMPTY_DATASOURCE_MAIN_TEXT,
 } from "@appsmith/constants/messages";
 import styled from "styled-components";
-import {
-  useDatasourceIdFromURL,
-  getExplorerStatus,
-  saveExplorerStatus,
-} from "@appsmith/pages/Editor/Explorer/helpers";
 import { Icon, Button } from "design-system";
 import { AddEntity, EmptyComponent } from "./common";
-import { integrationEditorURL } from "RouteBuilder";
 import { getCurrentAppWorkspace } from "@appsmith/selectors/workspaceSelectors";
 
 import type { AppState } from "@appsmith/reducers";
 import { DatasourceCreateEntryPoints } from "constants/Datasource";
-import AnalyticsUtil from "utils/AnalyticsUtil";
-import WalkthroughContext from "components/featureWalkthrough/walkthroughContext";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import {
   getHasCreateDatasourcePermission,
   getHasManageDatasourcePermission,
 } from "@appsmith/utils/BusinessFeatures/permissionPageHelpers";
+import {
+  useAppWideAndOtherDatasource,
+  useDatasourceSuggestions,
+} from "./hooks";
+import { getPlugins } from "@appsmith/selectors/entitiesSelector";
+import { keyBy } from "lodash";
+import { useDatasourceIdFromURL } from "@appsmith/pages/Editor/Explorer/helpers";
+import type { Datasource } from "entities/Datasource";
+import ExplorerDatasourceEntity from "./Datasources/DatasourceEntity";
+
+interface DatasourcesProps {
+  isDatasourcesOpen: boolean | null;
+  addDatasource: (source: string) => void;
+  onDatasourcesToggle: (isOpen: boolean) => void;
+  listDatasource: () => void;
+  entityId: string;
+}
 
 const ShowAllButton = styled(Button)`
   margin: 0.25rem 1.5rem;
 `;
 
-const Datasources = React.memo(() => {
+const Datasources = React.memo((props: DatasourcesProps) => {
   const { appWideDS, otherDS } = useAppWideAndOtherDatasource();
-  const pageId = useSelector(getCurrentPageId) || "";
-  const plugins = useSelector(getPlugins);
-  const applicationId = useSelector(getCurrentApplicationId);
-  const isDatasourcesOpen = getExplorerStatus(applicationId, "datasource");
-  const pluginGroups = React.useMemo(() => keyBy(plugins, "id"), [plugins]);
-  const { isOpened: isWalkthroughOpened, popFeature } =
-    useContext(WalkthroughContext) || {};
-
+  const {
+    addDatasource,
+    isDatasourcesOpen,
+    listDatasource,
+    onDatasourcesToggle,
+  } = props;
   const userWorkspacePermissions = useSelector(
     (state: AppState) => getCurrentAppWorkspace(state).userPermissions ?? [],
   );
-
+  const plugins = useSelector(getPlugins);
+  const pluginGroups = React.useMemo(() => keyBy(plugins, "id"), [plugins]);
   const isFeatureEnabled = useFeatureFlag(FEATURE_FLAG.license_gac_enabled);
 
   const canCreateDatasource = getHasCreateDatasourcePermission(
@@ -69,36 +63,8 @@ const Datasources = React.memo(() => {
     userWorkspacePermissions,
   );
 
-  const closeWalkthrough = useCallback(() => {
-    if (isWalkthroughOpened && popFeature) {
-      popFeature("EXPLORER_DATASOURCE_ADD");
-    }
-  }, [isWalkthroughOpened, popFeature]);
-
-  const addDatasource = (entryPoint: string) => {
-    history.push(
-      integrationEditorURL({
-        pageId,
-        selectedTab: INTEGRATION_TABS.NEW,
-      }),
-    );
-    // Event for datasource creation click
-    AnalyticsUtil.logEvent("NAVIGATE_TO_CREATE_NEW_DATASOURCE_PAGE", {
-      entryPoint,
-    });
-    closeWalkthrough();
-  };
   const activeDatasourceId = useDatasourceIdFromURL();
   const datasourceSuggestions = useDatasourceSuggestions();
-
-  const listDatasource = useCallback(() => {
-    history.push(
-      integrationEditorURL({
-        pageId,
-        selectedTab: INTEGRATION_TABS.ACTIVE,
-      }),
-    );
-  }, [pageId]);
 
   const datasourceElements = React.useMemo(
     () =>
@@ -113,23 +79,16 @@ const Datasources = React.memo(() => {
           <ExplorerDatasourceEntity
             canManageDatasource={canManageDatasource}
             datasource={datasource}
+            entityId={props.entityId}
             isActive={datasource.id === activeDatasourceId}
             key={datasource.id}
-            pageId={pageId}
             plugin={pluginGroups[datasource.pluginId]}
             searchKeyword={""}
             step={1}
           />
         );
       }),
-    [appWideDS, pageId, activeDatasourceId],
-  );
-
-  const onDatasourcesToggle = useCallback(
-    (isOpen: boolean) => {
-      saveExplorerStatus(applicationId, "datasource", isOpen);
-    },
-    [applicationId],
+    [appWideDS, props.entityId, activeDatasourceId],
   );
 
   return (
