+++ app/client/src/pages/Editor/Explorer/Datasources/DatasourceEntity.tsx
@@ -26,7 +26,6 @@ import {
   saasEditorDatasourceIdURL,
 } from "RouteBuilder";
 import { inGuidedTour } from "selectors/onboardingSelectors";
-import { getCurrentPageId } from "selectors/editorSelectors";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import { useLocation } from "react-router";
 import omit from "lodash/omit";
@@ -39,7 +38,7 @@ interface ExplorerDatasourceEntityProps {
   datasource: Datasource;
   step: number;
   searchKeyword?: string;
-  pageId: string;
+  entityId: string;
   isActive: boolean;
   canManageDatasource?: boolean;
 }
@@ -57,16 +56,16 @@ const DataStructureListWrapper = styled.div<{ height: number }>`
 
 const ExplorerDatasourceEntity = React.memo(
   (props: ExplorerDatasourceEntityProps) => {
+    const { entityId } = props;
     const guidedTourEnabled = useSelector(inGuidedTour);
     const dispatch = useDispatch();
-    const pageId = useSelector(getCurrentPageId);
     const icon = getPluginIcon(props.plugin);
     const location = useLocation();
     const switchDatasource = useCallback(() => {
       let url;
       if (props.plugin && props.plugin.type === PluginType.SAAS) {
         url = saasEditorDatasourceIdURL({
-          pageId,
+          pageId: entityId,
           pluginPackageName: props.plugin.packageName,
           datasourceId: props.datasource.id,
           params: {
@@ -75,7 +74,7 @@ const ExplorerDatasourceEntity = React.memo(
         });
       } else {
         url = datasourcesEditorIdURL({
-          pageId,
+          pageId: entityId,
           datasourceId: props.datasource.id,
           params: omit(getQueryParams(), "viewMode"),
         });
@@ -88,7 +87,12 @@ const ExplorerDatasourceEntity = React.memo(
         name: props.datasource.name,
       });
       history.push(url, { invokedBy: NavigationMethod.EntityExplorer });
-    }, [props.datasource.id, props.datasource.name, location.pathname, pageId]);
+    }, [
+      props.datasource.id,
+      props.datasource.name,
+      location.pathname,
+      entityId,
+    ]);
 
     const queryId = getQueryIdFromURL();
     const queryAction = useSelector((state: AppState) =>
