+++ app/client/src/pages/Editor/Explorer/EntityExplorer.test.tsx
@@ -57,6 +57,12 @@ jest
   .spyOn(explorerSelector, "getExplorerWidth")
   .mockImplementation(() => DEFAULT_ENTITY_EXPLORER_WIDTH);
 
+const addDatasource = jest.fn();
+const listDatasource = jest.fn();
+const onDatasourcesToggle = jest.fn();
+const isDatasourcesOpen = true;
+const entityId = "pageId";
+
 describe("Entity Explorer tests", () => {
   beforeAll(() => {
     runSagaMiddleware();
@@ -89,7 +95,15 @@ describe("Entity Explorer tests", () => {
       .spyOn(permissionPageHelpers, "getHasCreateDatasourcePermission")
       .mockReturnValue(true);
     store.dispatch(updateCurrentPage("pageId"));
-    const component = render(<Datasources />);
+    const component = render(
+      <Datasources
+        addDatasource={addDatasource}
+        entityId={entityId}
+        isDatasourcesOpen={isDatasourcesOpen}
+        listDatasource={listDatasource}
+        onDatasourcesToggle={onDatasourcesToggle}
+      />,
+    );
     expect(component.container.getElementsByClassName("t--entity").length).toBe(
       5,
     );
@@ -105,7 +119,15 @@ describe("Entity Explorer tests", () => {
     const mockExplorerState = jest.spyOn(helpers, "getExplorerStatus");
     mockExplorerState.mockImplementationOnce(() => true);
     store.dispatch(updateCurrentPage("pageId"));
-    const component = render(<Datasources />);
+    const component = render(
+      <Datasources
+        addDatasource={addDatasource}
+        entityId={entityId}
+        isDatasourcesOpen={isDatasourcesOpen}
+        listDatasource={listDatasource}
+        onDatasourcesToggle={onDatasourcesToggle}
+      />,
+    );
     expect(component.container.getElementsByClassName("t--entity").length).toBe(
       4,
     );
@@ -113,37 +135,6 @@ describe("Entity Explorer tests", () => {
       "entity-add_new_datasource",
     );
     expect(addDatasourceEntity).toBeNull();
-  });
-  it("should hide delete & edit of datasource if the user don't have valid permissions", async () => {
-    store.dispatch({
-      type: ReduxActionTypes.FETCH_DATASOURCES_SUCCESS,
-      payload: mockDatasources,
-    });
-    jest
-      .spyOn(permissionPageHelpers, "getHasCreateDatasourcePermission")
-      .mockReturnValue(true);
-    jest
-      .spyOn(permissionPageHelpers, "getHasManageDatasourcePermission")
-      .mockReturnValue(false);
-    jest
-      .spyOn(permissionPageHelpers, "getHasDeleteDatasourcePermission")
-      .mockReturnValue(false);
-    const mockExplorerState = jest.spyOn(helpers, "getExplorerStatus");
-    mockExplorerState.mockImplementationOnce(() => true);
-    store.dispatch(updateCurrentPage("pageId"));
-    const { container } = render(<Datasources />);
-    const target = container.getElementsByClassName("t--context-menu");
-    await userEvent.click(target[2]);
-    const deleteOption = document.getElementsByClassName(
-      "t--datasource-delete",
-    );
-    const editOption = document.getElementsByClassName("t--datasource-rename");
-    const refreshOption = document.getElementsByClassName(
-      "t--datasource-refresh",
-    );
-    expect(deleteOption.length).toBe(0);
-    expect(editOption.length).toBe(0);
-    expect(refreshOption.length).toBe(1);
   });
   it("Should render Widgets tree in entity explorer", () => {
     const children: any = buildChildren([{ type: "TABS_WIDGET" }]);
@@ -409,5 +400,45 @@ describe("Entity Explorer tests", () => {
         undefined,
       );
     });
+  });
+
+  it("should hide delete & edit of datasource if the user don't have valid permissions", async () => {
+    store.dispatch({
+      type: ReduxActionTypes.FETCH_DATASOURCES_SUCCESS,
+      payload: mockDatasources,
+    });
+    jest
+      .spyOn(permissionPageHelpers, "getHasCreateDatasourcePermission")
+      .mockReturnValue(true);
+    jest
+      .spyOn(permissionPageHelpers, "getHasManageDatasourcePermission")
+      .mockReturnValue(false);
+    jest
+      .spyOn(permissionPageHelpers, "getHasDeleteDatasourcePermission")
+      .mockReturnValue(false);
+    const mockExplorerState = jest.spyOn(helpers, "getExplorerStatus");
+    mockExplorerState.mockImplementationOnce(() => true);
+    store.dispatch(updateCurrentPage("pageId"));
+    const { container } = render(
+      <Datasources
+        addDatasource={addDatasource}
+        entityId={entityId}
+        isDatasourcesOpen={isDatasourcesOpen}
+        listDatasource={listDatasource}
+        onDatasourcesToggle={onDatasourcesToggle}
+      />,
+    );
+    const target = container.getElementsByClassName("t--context-menu");
+    await userEvent.click(target[2]);
+    const deleteOption = document.getElementsByClassName(
+      "t--datasource-delete",
+    );
+    const editOption = document.getElementsByClassName("t--datasource-rename");
+    const refreshOption = document.getElementsByClassName(
+      "t--datasource-refresh",
+    );
+    expect(deleteOption.length).toBe(0);
+    expect(editOption.length).toBe(0);
+    expect(refreshOption.length).toBe(1);
   });
 });
