+++ app/client/src/pages/Editor/Explorer/EntityExplorer.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useCallback, useEffect } from "react";
+import React, { useRef, useCallback, useEffect, useContext } from "react";
 import styled from "styled-components";
 import { NonIdealState, Classes } from "@blueprintjs/core";
 import JSDependencies from "./Libraries";
@@ -23,16 +23,19 @@ import { fetchWorkspace } from "@appsmith/actions/workspaceActions";
 import { getCurrentWorkspaceId } from "@appsmith/selectors/workspaceSelectors";
 import { importSvg } from "design-system-old";
 import AnalyticsUtil from "utils/AnalyticsUtil";
+import { EntityExplorerWrapper } from "./Common/EntityExplorerWrapper";
+import { getCurrentApplicationId } from "selectors/editorSelectors";
+import { INTEGRATION_TABS } from "constants/routes";
+import {
+  getExplorerStatus,
+  saveExplorerStatus,
+} from "@appsmith/pages/Editor/Explorer/helpers";
+import { integrationEditorURL } from "RouteBuilder";
+import WalkthroughContext from "components/featureWalkthrough/walkthroughContext";
 
 const NoEntityFoundSvg = importSvg(
   async () => import("assets/svg/no_entities_found.svg"),
 );
-
-const Wrapper = styled.div`
-  height: 100%;
-  overflow-y: auto;
-  -ms-overflow-style: none;
-`;
 
 const NoResult = styled(NonIdealState)`
   &.${Classes.NON_IDEAL_STATE} {
@@ -87,13 +90,52 @@ function EntityExplorer({ isActive }: { isActive: boolean }) {
     dispatch(fetchWorkspace(currentWorkspaceId));
   }, [currentWorkspaceId]);
 
+  const { isOpened: isWalkthroughOpened, popFeature } =
+    useContext(WalkthroughContext) || {};
+  const applicationId = useSelector(getCurrentApplicationId);
+  const isDatasourcesOpen = getExplorerStatus(applicationId, "datasource");
+
+  const closeWalkthrough = useCallback(() => {
+    if (isWalkthroughOpened && popFeature) {
+      popFeature("EXPLORER_DATASOURCE_ADD");
+    }
+  }, [isWalkthroughOpened, popFeature]);
+
+  const addDatasource = useCallback(
+    (entryPoint: string) => {
+      history.push(
+        integrationEditorURL({
+          pageId,
+          selectedTab: INTEGRATION_TABS.NEW,
+        }),
+      );
+      // Event for datasource creation click
+      AnalyticsUtil.logEvent("NAVIGATE_TO_CREATE_NEW_DATASOURCE_PAGE", {
+        entryPoint,
+      });
+      closeWalkthrough();
+    },
+    [pageId, closeWalkthrough],
+  );
+
+  const listDatasource = useCallback(() => {
+    history.push(
+      integrationEditorURL({
+        pageId,
+        selectedTab: INTEGRATION_TABS.ACTIVE,
+      }),
+    );
+  }, [pageId]);
+
+  const onDatasourcesToggle = useCallback(
+    (isOpen: boolean) => {
+      saveExplorerStatus(applicationId, "datasource", isOpen);
+    },
+    [applicationId],
+  );
+
   return (
-    <Wrapper
-      className={`t--entity-explorer-wrapper relative overflow-y-auto ${
-        isActive ? "" : "hidden"
-      }`}
-      ref={explorerRef}
-    >
+    <EntityExplorerWrapper explorerRef={explorerRef} isActive={isActive}>
       <ExplorerWidgetGroup
         addWidgetsFn={showWidgetsSidebar}
         searchKeyword=""
@@ -108,9 +150,15 @@ function EntityExplorer({ isActive }: { isActive: boolean }) {
           title="No entities found"
         />
       )}
-      <Datasources />
+      <Datasources
+        addDatasource={addDatasource}
+        entityId={pageId}
+        isDatasourcesOpen={isDatasourcesOpen}
+        listDatasource={listDatasource}
+        onDatasourcesToggle={onDatasourcesToggle}
+      />
       <JSDependencies />
-    </Wrapper>
+    </EntityExplorerWrapper>
   );
 }
