+++ app/client/src/pages/Editor/index.tsx
@@ -155,7 +155,7 @@ class Editor extends Component<Props> {
           <Helmet>
             <meta charSet="utf-8" />
             <title>
-              {`${this.props.currentApplicationName} |`} Editor | Appsmith
+              {`${this.props.currentApplicationName} | Editor | Appsmith`}
             </title>
           </Helmet>
           <GlobalHotKeys>
